Avoid useless SELECT queries on the first synchronization.
Making SELECT queries while performing the first synchronization is not
necessary since the tables are supposed to be empty. Making these SELECT was
affecting performance so this commit provides a way to force queries to be
INSERT ones and by doing so it avoids useless SELECT.

Set default last_sync value to 0 to avoid synchronization cancelled if the
first one fails.

Simplify functions a bit by calling a generic function to execute insert or
update queries.

Add indexes to database.