Handle deletes and misses due to caching.
Deletes were never handled, so there's now a quick check for records
with the status of "deleted" which should be purged from the database.
Ideally this would be done when the update phase is happening to
needlessly write the deleted entry first, but it would be quite
difficult due to the way that code is written... So this works!

When it comes to caching, a query to PeeringDB goes directly through the
application and to the database, except when you request all the records
(the "since" parameter is set to zero) when for reasons of lower load it
serves up a stored copy that may be up to 15 minutes old.

Rather than storing the time the last sync was completed (and therefore
potentially missing some records that may have been added / changed
since the cache was generated) we now just look for the timestamp of the
last record to be updated, and tell PeeringDB to give us everything
since then.

This has the unfortunate effect of needlessly performing an insert /
delete cycle if the only fetched records are deleted, but I can live
with that for now!